fix change pass
