+++ FE/src/API/Endpoints/authEndpoint.ts
@@ -1,4 +1,4 @@
-import { PasswordSend, UserSend } from '../DTOs/userTypes'
+import { UserSend } from '../DTOs/userTypes'
 import axios from 'axios'
 
 export const registerUser = async (user: UserSend) => {
@@ -24,22 +24,6 @@ export const logIn = async (user: UserSend) => {
     method: 'post',
     maxBodyLength: Infinity,
     url: process.env.REACT_APP_DB_URL + '/auth/login',
-    headers: { 
-      'Content-Type': 'application/json'
-    },
-    data : data
-  };
-
-  return axios.request(config).then((response) => response.data)
-}
-
-export const changePassword = async (user: PasswordSend) => {
-  let data = JSON.stringify(user);
-  console.log(data)
-  let config = {
-    method: 'put',
-    maxBodyLength: Infinity,
-    url: process.env.REACT_APP_DB_URL + '/users/change-password',
     headers: { 
       'Content-Type': 'application/json'
     },
